Navigation Drawer Created.
MainActivity:
added the fragment_navigation_drawer onto my modified setup function.

NavigationDrawerFragment:
Created two methods one to save preferance settings and the other to
read from the shared preference values.

Added two boolean variables to determined if it is the first time the
user has opened the drawer or not.

Created a DrawerLayout post with a Runnable  override thread for the run
function which syncs the states of the drawer.

activity_main:
Changed the width and changed gravity for layout_gravity which fix the
fragment layout from laying on top of the activity_main.

fragment_navigation_drawer:
added a white background color to the navigation drawer.