Add Ingredients Layout
create a add ingredient activity added the components needed.

Need to look up how spinners work in order to make them dynamically
through java in the future. For now I need to know how to do it in xml.

for activity_main and activity_sub they are being deprecated since
everything is located in activity_main_appbar but i am keeping it for
reference. I put some comments to remind myself to delete them for
future builds of the app.