Implemented Refrigerator and FoodItem classes
AddIngredients:
You can now change to the PopUpForm activity from any button and I added
a string to change the header of the form to notify the user which
section he is on.

FoodItem:
This class is a container for food attributes.

PopUpForm:
Storing the user input in the FoodItem Container and then passing that
container the Refrigerator class alongside the header of the form to
know what type the food will belong to on the Refrigerator.

Refrigerator class:
The purpose of this class is to hold all of your food items and put them
in sub-categories. The sub-categories are maps. Example proteinsDic<food
name, food object>
There is a dictionary for each category.

Deleted SubActivity and activity_sub.xml

changed some variable names in popupform.xml and added some strings that
will be displayed to the used on the strings.xml.