Changing to enums
AddIngredients:
Move Initialization to their own methods because it was getting hard to
read the onCreate method. Plus everything is separated so it is easier
to take out or add stuff.

PopUpForm:
added class method call addToSpinners to add the new food item to the
spinner.

Refrigerator:
Learned a lot about enums in java and if everything works they ways it
should be then we are switching to enums since it is clearer and easier
to read.

testing on next commit to decide which version to keep.