Refactor,Fixed Spinners, Added Lots more food groups
Android Manifest:
Nothing. But just to be safe I'll add it since it tracked a change and
it is working perfectly.

AddIngredients:
- Refactor, Added New Buttons, Added New Spinners
All according to the new Food Group in the enum.
- Learned and added Start, Resume, Pause, states since it was deleting
my data when pausing the activity.  Solved by saving some of my data
onPause.
- TODO 0: Get rid of singleton on AddIngredients, there is no need
anymore.
- TODO 1: Use SaveInstanceData and RestoreInstanceData to save the
spinner used added data.

PopUpForm:
- Refactor
- Deleted add ingredients.getInstance and used the Refrigerator as the
main communication from PopUpForm to AddIngredients by using the newly
add setters and getters that keep track of the latest foodItem and
FoodType added onto the refrigerator.

Refrigerator:
- Refactored
- Added New Setters and Getters for the latest food type and foodItem.
- Changed CONFECTIONS constant enum class to SUGARS, something more
understandable to the user.
- Changed the logs.d to logs.i to better keep track of it.

Compiler Changes:
- Added new filters for each and every class.
- TODO: Add a filter that tracks all the classes.

add_ingredients.xml:
- Added all the foodgroup spinners
- changed all their ids to keep consistency
- refactor
- added a scrollview
- TODO: Debating wether to use a button to go to the top of the scroll
view. It is not that long of a swipe to reach the bottom of screen.
However I will make the margin spaces bigger so that might change.
- TODO: Change Marging Spaces to be larger between each spinner.
- TODO: Make the spinners shorter, do not "match parent" for height make
custom.

strings.xml:
- Removed static array list for a more dynamic arraylist on
addIngredients.java file.
- Added All the headers for the group types.